filter and sort the packages

+++ src/main/java/com/mobiquityinc/entities/PackageVO.java
@@ -4,16 +4,16 @@ import java.util.List;
 
 public class PackageVO {
 
-    private Integer weightPackage;
+    private Double weightPackage;
 
     private List<ThingVO> things;
 
-    public PackageVO(Integer weightPackage, List<ThingVO> things) {
+    public PackageVO(Double weightPackage, List<ThingVO> things) {
         this.weightPackage = weightPackage;
         this.things = things;
     }
 
-    public Integer getWeightPackage() {
+    public Double getWeightPackage() {
         return weightPackage;
     }
